rework for settings tests and loading of scenarios
