+++ e2e/smoke_test/README.md
@@ -1,0 +1,16 @@
+
+
+Loop Functionality Smoketest
+    [] Pair Pod
+    [] Connect to Dexcom G6
+    [] Add Carbs
+    [] Deliver bolus
+    [] View Carbohydrates
+    [] Set Workout mode
+    [] Set pre-meal override
+    [] View pod status
+    [] View Settings
+    [] Verify closed loop basal changes
+
+
+https://docs.google.com/document/d/1fz-3zoMSmnvHQBw4fFmm64tY--VCoKUb2DGvh7ygIaI/edit#
