+++ src/index.js
@@ -1,0 +1,8 @@
+const  setup = require('./setup');
+const  {match, assert} = require('./loopUI');
+
+module.exports = {
+    match,
+    assert,
+    setup
+};
